+++ Mul-Ty-Player Client/Classes/GamePlay/LevelHandler.cs
@@ -53,6 +53,8 @@ internal class LevelHandler
             HSync.SyncObjects["TE"].GlobalObjectData.ContainsKey(CurrentLevelId) &&
             (HSync.SyncObjects["TE"].SaveSync as SaveTESyncer)?.GlobalSaveData[CurrentLevelId][3] == 1)
             (HSync.SyncObjects["TE"] as TEHandler)?.ShowStopwatch();
+        if (SettingsHandler.DoMushroomSyncing)
+            MushroomHandler.CheckMushrooms();
         if (CurrentLevelData.Id != 16)
         {
             HKoala.SetBaseAddress();